+++ CHANGELOG.md
@@ -3,6 +3,138 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.15.0](https://github.com/wework/ray/compare/v1.14.0...v1.15.0) (2019-11-01)
+
+
+### Features
+
+* **core-react:** change "type" to "variant" on Button ([#241](https://github.com/wework/ray/issues/241)) ([1d47fd4](https://github.com/wework/ray/commit/1d47fd4))
+
+
+
+
+
+# [1.14.0](https://github.com/wework/ray/compare/v1.13.10...v1.14.0) (2019-11-01)
+
+
+### Features
+
+* **core:** add support for append add-on which similar to prepend ([#246](https://github.com/wework/ray/issues/246)) ([60f5de2](https://github.com/wework/ray/commit/60f5de2))
+
+
+
+
+
+## [1.13.10](https://github.com/wework/ray/compare/v1.13.9...v1.13.10) (2019-10-30)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.9](https://github.com/wework/ray/compare/v1.13.8...v1.13.9) (2019-10-29)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.8](https://github.com/wework/ray/compare/v1.13.7...v1.13.8) (2019-10-28)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.7](https://github.com/wework/ray/compare/v1.13.6...v1.13.7) (2019-10-25)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.6](https://github.com/wework/ray/compare/v1.13.5...v1.13.6) (2019-10-24)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.5](https://github.com/wework/ray/compare/v1.13.4...v1.13.5) (2019-10-21)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.4](https://github.com/wework/ray/compare/v1.13.3...v1.13.4) (2019-10-15)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.3](https://github.com/wework/ray/compare/v1.13.2...v1.13.3) (2019-10-14)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.2](https://github.com/wework/ray/compare/v1.13.1...v1.13.2) (2019-10-10)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.13.1](https://github.com/wework/ray/compare/v1.13.0...v1.13.1) (2019-10-08)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+# [1.13.0](https://github.com/wework/ray/compare/v1.12.4...v1.13.0) (2019-10-01)
+
+
+### Features
+
+* **grid:** add product grid to ray product ([#234](https://github.com/wework/ray/issues/234)) ([7d8b4c9](https://github.com/wework/ray/commit/7d8b4c9))
+
+
+
+
+
+## [1.12.4](https://github.com/wework/ray/compare/v1.12.3...v1.12.4) (2019-09-18)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [1.12.3](https://github.com/wework/ray/compare/v1.12.2...v1.12.3) (2019-09-17)
+
+
+### Bug Fixes
+
+* **chip:** add disabled state ([#231](https://github.com/wework/ray/issues/231)) ([d4230db](https://github.com/wework/ray/commit/d4230db))
+
+
+
+
+
 ## [1.12.2](https://github.com/wework/ray/compare/v1.12.1...v1.12.2) (2019-09-16)
 
 **Note:** Version bump only for package root