+++ packages/core-react/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.15.0](https://github.com/wework/ray/compare/v1.14.0...v1.15.0) (2019-11-01)
+
+
+### Features
+
+* **core-react:** change "type" to "variant" on Button ([#241](https://github.com/wework/ray/issues/241)) ([1d47fd4](https://github.com/wework/ray/commit/1d47fd4))
+
+
+
+
+
+# [1.14.0](https://github.com/wework/ray/compare/v1.13.10...v1.14.0) (2019-11-01)
+
+**Note:** Version bump only for package @wework/ray-core-react
+
+
+
+
+
+## [1.13.8](https://github.com/wework/ray/compare/v1.13.7...v1.13.8) (2019-10-28)
+
+**Note:** Version bump only for package @wework/ray-core-react
+
+
+
+
+
+## [1.13.7](https://github.com/wework/ray/compare/v1.13.6...v1.13.7) (2019-10-25)
+
+**Note:** Version bump only for package @wework/ray-core-react
+
+
+
+
+
+## [1.12.4](https://github.com/wework/ray/compare/v1.12.3...v1.12.4) (2019-09-18)
+
+**Note:** Version bump only for package @wework/ray-core-react
+
+
+
+
+
+## [1.12.3](https://github.com/wework/ray/compare/v1.12.2...v1.12.3) (2019-09-17)
+
+**Note:** Version bump only for package @wework/ray-core-react
+
+
+
+
+
 ## [1.12.2](https://github.com/wework/ray/compare/v1.12.1...v1.12.2) (2019-09-16)
 
 **Note:** Version bump only for package @wework/ray-core-react