+++ packages/docs-app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.14.0](https://github.com/wework/ray/compare/v1.13.10...v1.14.0) (2019-11-01)
+
+**Note:** Version bump only for package ray-website-gatsby
+
+
+
+
+
+## [1.12.3](https://github.com/wework/ray/compare/v1.12.2...v1.12.3) (2019-09-17)
+
+**Note:** Version bump only for package ray-website-gatsby
+
+
+
+
+
 ## [1.12.2](https://github.com/wework/ray/compare/v1.12.1...v1.12.2) (2019-09-16)
 
 **Note:** Version bump only for package ray-website-gatsby